Inline ValidationBuilder
