+++ src/commonMain/kotlin/io/konform/validation/ValidationBuilder.kt
@@ -1,11 +1,18 @@
 package io.konform.validation
 
+import io.konform.validation.ValidationBuilder.PropModifier.NonNull
+import io.konform.validation.ValidationBuilder.PropModifier.Optional
+import io.konform.validation.ValidationBuilder.PropModifier.OptionalRequired
 import io.konform.validation.internal.ArrayValidation
 import io.konform.validation.internal.IterableValidation
 import io.konform.validation.internal.MapValidation
+import io.konform.validation.internal.NonNullPropertyValidation
+import io.konform.validation.internal.OptionalPropertyValidation
 import io.konform.validation.internal.OptionalValidation
+import io.konform.validation.internal.RequiredPropertyValidation
 import io.konform.validation.internal.RequiredValidation
-import io.konform.validation.internal.ValidationBuilderImpl
+import io.konform.validation.internal.ValidationNode
+import io.konform.validation.kotlin.Grammar
 import kotlin.jvm.JvmName
 import kotlin.reflect.KFunction1
 import kotlin.reflect.KProperty1
@@ -14,34 +21,53 @@ import kotlin.reflect.KProperty1
 private annotation class ValidationScope
 
 @ValidationScope
-public abstract class ValidationBuilder<T> {
-    public abstract fun build(): Validation<T>
+public class ValidationBuilder<T> {
+    public fun build(): Validation<T> {
+        val nestedValidations =
+            subValidations.map { (key, builder) ->
+                key.build(builder)
+            }
+        return ValidationNode(constraints, nestedValidations + prebuiltValidations)
+    }
 
-    public abstract fun addConstraint(
+    public fun addConstraint(
         errorMessage: String,
         vararg templateValues: String,
         test: (T) -> Boolean,
-    ): Constraint<T>
+    ): Constraint<T> = Constraint(errorMessage, templateValues.toList(), test).also { constraints.add(it) }
 
-    public abstract infix fun Constraint<T>.hint(hint: String): Constraint<T>
+    public fun Constraint<T>.hint(hint: String): Constraint<T> =
+        Constraint(hint, this.templateValues, this.test).also {
+            constraints.remove(this)
+            constraints.add(it)
+        }
 
-    internal abstract fun <R> onEachIterable(
+    public fun <R> onEachIterable(
         name: String,
         prop: (T) -> Iterable<R>,
         init: ValidationBuilder<R>.() -> Unit,
-    )
+    ) {
+        requireValidName(name)
+        init(prop.getOrCreateIterablePropertyBuilder(name, NonNull))
+    }
 
-    internal abstract fun <R> onEachArray(
+    public fun <R> onEachArray(
         name: String,
         prop: (T) -> Array<R>,
         init: ValidationBuilder<R>.() -> Unit,
-    )
+    ) {
+        requireValidName(name)
+        init(ArrayPropKey(prop, name, NonNull).getOrCreateBuilder())
+    }
 
-    internal abstract fun <K, V> onEachMap(
+    public fun <K, V> onEachMap(
         name: String,
         prop: (T) -> Map<K, V>,
         init: ValidationBuilder<Map.Entry<K, V>>.() -> Unit,
-    )
+    ) {
+        requireValidName(name)
+        init(MapPropKey(prop, name, NonNull).getOrCreateBuilder())
+    }
 
     @JvmName("onEachIterable")
     public infix fun <R> KProperty1<T, Iterable<R>>.onEach(init: ValidationBuilder<R>.() -> Unit): Unit = onEachIterable(name, this, init)
@@ -76,48 +102,146 @@ public abstract class ValidationBuilder<T> {
 
     public infix fun <R> KFunction1<T, R?>.required(init: ValidationBuilder<R>.() -> Unit): Unit = required("$name()", this, init)
 
-    /**
-     * Calculate a value from the input and run a validation on it.
-     * @param name The name that should be reported in validation errors. Must be a valid kotlin name, optionally followed by ().
-     * @param f The function for which you want to validate the result of
-     * @see run
-     */
-    public abstract fun <R> validate(
+    public fun <R> validate(
         name: String,
         f: (T) -> R,
         init: ValidationBuilder<R>.() -> Unit,
-    )
+    ): Unit = init(f.getOrCreateBuilder(name, NonNull))
 
-    /**
-     * Calculate a value from the input and run a validation on it, but only if the value is not null.
-     */
-    public abstract fun <R> ifPresent(
+    public fun <R> ifPresent(
         name: String,
         f: (T) -> R?,
         init: ValidationBuilder<R>.() -> Unit,
-    )
+    ): Unit = init(f.getOrCreateBuilder(name, Optional))
 
-    /**
-     * Calculate a value from the input and run a validation on it, and give an error if the result is null.
-     */
-    public abstract fun <R> required(
+    public fun <R> required(
         name: String,
         f: (T) -> R?,
         init: ValidationBuilder<R>.() -> Unit,
-    )
+    ): Unit = init(f.getOrCreateBuilder(name, OptionalRequired))
 
-    /** Run an arbitrary other validation. */
-    public abstract fun run(validation: Validation<T>)
+    public fun run(validation: Validation<T>) {
+        prebuiltValidations.add(validation)
+    }
 
-    public abstract val <R> KProperty1<T, R>.has: ValidationBuilder<R>
-    public abstract val <R> KFunction1<T, R>.has: ValidationBuilder<R>
+    private fun <R> ((T) -> R?).getOrCreateBuilder(
+        name: String,
+        modifier: PropModifier,
+    ): ValidationBuilder<R> {
+        requireValidName(name)
+        val key = SingleValuePropKey(this, name, modifier)
+        @Suppress("UNCHECKED_CAST")
+        return (subValidations.getOrPut(key) { ValidationBuilder<R>() } as ValidationBuilder<R>)
+    }
+
+    private fun <R> ((T) -> Iterable<R>).getOrCreateIterablePropertyBuilder(
+        name: kotlin.String,
+        modifier: PropModifier,
+    ): ValidationBuilder<R> {
+        val key = IterablePropKey(this, name, modifier)
+        @Suppress("UNCHECKED_CAST")
+        return (subValidations.getOrPut(key) { ValidationBuilder<R>() } as ValidationBuilder<R>)
+    }
+
+    private fun <R> PropKey<T>.getOrCreateBuilder(): ValidationBuilder<R> {
+        @Suppress("UNCHECKED_CAST")
+        return (subValidations.getOrPut(this) { ValidationBuilder<R>() } as ValidationBuilder<R>)
+    }
+
+    private fun requireValidName(name: String) =
+        require(Grammar.Identifier.isValid(name) || Grammar.FunctionDeclaration.isUnary(name)) {
+            "'$name' is not a valid kotlin identifier or getter name."
+        }
+
+    public val <R> KProperty1<T, R>.has: ValidationBuilder<R>
+        get() = getOrCreateBuilder(name, NonNull)
+    public val <R> KFunction1<T, R>.has: ValidationBuilder<R>
+        get() = getOrCreateBuilder(name, NonNull)
+
+    private val constraints = mutableListOf<Constraint<T>>()
+    private val subValidations = mutableMapOf<PropKey<T>, ValidationBuilder<*>>()
+    private val prebuiltValidations = mutableListOf<Validation<T>>()
+
+    private enum class PropModifier {
+        NonNull,
+        Optional,
+        OptionalRequired,
+    }
+
+    private abstract class PropKey<T> {
+        abstract fun build(builder: ValidationBuilder<*>): Validation<T>
+    }
+
+    private data class SingleValuePropKey<T, R>(
+        val property: (T) -> R,
+        val name: String,
+        val modifier: PropModifier,
+    ) : PropKey<T>() {
+        override fun build(builder: ValidationBuilder<*>): Validation<T> {
+            @Suppress("UNCHECKED_CAST")
+            val validations = (builder as ValidationBuilder<R>).build()
+            return when (modifier) {
+                NonNull -> NonNullPropertyValidation(property, name, validations)
+                Optional -> OptionalPropertyValidation(property, name, validations)
+                OptionalRequired -> RequiredPropertyValidation(property, name, validations)
+            }
+        }
+    }
+
+    private data class IterablePropKey<T, R>(
+        val property: (T) -> Iterable<R>,
+        val name: String,
+        val modifier: PropModifier,
+    ) : PropKey<T>() {
+        override fun build(builder: ValidationBuilder<*>): Validation<T> {
+            @Suppress("UNCHECKED_CAST")
+            val validations = (builder as ValidationBuilder<R>).build()
+            return when (modifier) {
+                NonNull -> NonNullPropertyValidation(property, name, IterableValidation(validations))
+                Optional -> OptionalPropertyValidation(property, name, IterableValidation(validations))
+                OptionalRequired -> RequiredPropertyValidation(property, name, IterableValidation(validations))
+            }
+        }
+    }
+
+    private data class ArrayPropKey<T, R>(
+        val property: (T) -> Array<R>,
+        val name: String,
+        val modifier: PropModifier,
+    ) : PropKey<T>() {
+        override fun build(builder: ValidationBuilder<*>): Validation<T> {
+            @Suppress("UNCHECKED_CAST")
+            val validations = (builder as ValidationBuilder<R>).build()
+            return when (modifier) {
+                NonNull -> NonNullPropertyValidation(property, name, ArrayValidation(validations))
+                Optional -> OptionalPropertyValidation(property, name, ArrayValidation(validations))
+                OptionalRequired -> RequiredPropertyValidation(property, name, ArrayValidation(validations))
+            }
+        }
+    }
+
+    private data class MapPropKey<T, K, V>(
+        val property: (T) -> Map<K, V>,
+        val name: String,
+        val modifier: PropModifier,
+    ) : PropKey<T>() {
+        override fun build(builder: ValidationBuilder<*>): Validation<T> {
+            @Suppress("UNCHECKED_CAST")
+            val validations = (builder as ValidationBuilder<Map.Entry<K, V>>).build()
+            return when (modifier) {
+                NonNull -> NonNullPropertyValidation(property, name, MapValidation(validations))
+                Optional -> OptionalPropertyValidation(property, name, MapValidation(validations))
+                OptionalRequired -> RequiredPropertyValidation(property, name, MapValidation(validations))
+            }
+        }
+    }
 }
 
 /**
  * Run a validation if the property is not-null, and allow nulls.
  */
 public fun <T : Any> ValidationBuilder<T?>.ifPresent(init: ValidationBuilder<T>.() -> Unit) {
-    val builder = ValidationBuilderImpl<T>()
+    val builder = ValidationBuilder<T>()
     init(builder)
     run(OptionalValidation(builder.build()))
 }
@@ -126,14 +250,14 @@ public fun <T : Any> ValidationBuilder<T?>.ifPresent(init: ValidationBuilder<T>.
  * Run a validation on a nullable property, giving an error on nulls.
  */
 public fun <T : Any> ValidationBuilder<T?>.required(init: ValidationBuilder<T>.() -> Unit) {
-    val builder = ValidationBuilderImpl<T>()
+    val builder = ValidationBuilder<T>()
     init(builder)
     run(RequiredValidation(builder.build()))
 }
 
 @JvmName("onEachIterable")
 public fun <S, T : Iterable<S>> ValidationBuilder<T>.onEach(init: ValidationBuilder<S>.() -> Unit) {
-    val builder = ValidationBuilderImpl<S>()
+    val builder = ValidationBuilder<S>()
     init(builder)
     @Suppress("UNCHECKED_CAST")
     run(IterableValidation(builder.build()) as Validation<T>)
@@ -141,14 +265,14 @@ public fun <S, T : Iterable<S>> ValidationBuilder<T>.onEach(init: ValidationBuil
 
 @JvmName("onEachArray")
 public fun <T> ValidationBuilder<Array<T>>.onEach(init: ValidationBuilder<T>.() -> Unit) {
-    val builder = ValidationBuilderImpl<T>()
+    val builder = ValidationBuilder<T>()
     init(builder)
     run(ArrayValidation(builder.build()))
 }
 
 @JvmName("onEachMap")
 public fun <K, V, T : Map<K, V>> ValidationBuilder<T>.onEach(init: ValidationBuilder<Map.Entry<K, V>>.() -> Unit) {
-    val builder = ValidationBuilderImpl<Map.Entry<K, V>>()
+    val builder = ValidationBuilder<Map.Entry<K, V>>()
     init(builder)
     @Suppress("UNCHECKED_CAST")
     run(MapValidation(builder.build()) as Validation<T>)
